* Fix dynamics: 	- by using correct rotation for transforming the torque from world space to body reference frame of the inertia matrix 	- by calculating correct intertia matrix through correcting volume integrals prior to calculating the inertia tensor * Optimize/Simplify voxelization: Avoid switching of x and z-axis during voxelization by applying the view correction transformation only to the projected position in the vertex shader. This way, the coordinates for calculating the voxelization remain in the default coordinates which simplifies the shader code, because it avoids the switching of x and z coordinates. The principle stays the same * Perform conservative voxelization is once from all 3 global axes to reach a better voxel coverage * Remove compiler warning in dx11widget

+++ WindSim/src/GUI/dx11widget.cpp
@@ -25,7 +25,7 @@ DX11Widget::DX11Widget(QWidget* parent, Qt::WindowFlags flags)
 	if (Simulator::initOpenCLNecessary())
 	{
 		static auto info = wtl::getOpenCLInfo();
-		if (info.size() == 0 || std::accumulate(info.begin(), info.end(), 0, [](size_t v, const std::pair<wtl::description, std::vector<wtl::description>>& e) { return v + e.second.size(); }) == 0)
+		if (info.size() == 0 || std::accumulate(info.begin(), info.end(), 0, [](int v, const std::pair<wtl::description, std::vector<wtl::description>>& e) { return v + static_cast<int>(e.second.size()); }) == 0)
 		{
 			QMessageBox::critical(parent, "OpenCL Error", "No OpenCL platforms and devices available! Install the necessary SDK's by Intel or AMD!");
 			throw std::runtime_error("ERROR: Failed to initialize OpenCL!");
